lol value class
