+++ rsc/src/main/scala/rsc/semantics/Symbols.scala
@@ -3,8 +3,10 @@
 package rsc.semantics
 
 import rsc.gensym._
+import scala.collection.mutable
+import scala.compat.Platform.EOL
 import scala.meta.internal.{semanticdb => s}
-import scala.meta.internal.semanticdb.Scala.{Symbols, Descriptor => D}
+import scala.meta.internal.semanticdb.Scala.{Symbols, Descriptor => d}
 
 trait Symbols {
   type Symbol = String
@@ -22,27 +24,27 @@ trait Symbols {
   }
 
   def TermSymbol(owner: Symbol, value: String): Symbol = {
-    s.Scala.Symbols.Global(owner, D.Term(value))
+    s.Scala.Symbols.Global(owner, d.Term(value))
   }
 
   def MethodSymbol(owner: Symbol, value: String, disambig: String): Symbol = {
-    s.Scala.Symbols.Global(owner, D.Method(value, disambig))
+    s.Scala.Symbols.Global(owner, d.Method(value, disambig))
   }
 
   def TypeSymbol(owner: Symbol, value: String): Symbol = {
-    s.Scala.Symbols.Global(owner, D.Type(value))
+    s.Scala.Symbols.Global(owner, d.Type(value))
   }
 
   def PackageSymbol(owner: Symbol, value: String): Symbol = {
-    s.Scala.Symbols.Global(owner, D.Package(value))
+    s.Scala.Symbols.Global(owner, d.Package(value))
   }
 
   def ParamSymbol(owner: Symbol, value: String): Symbol = {
-    s.Scala.Symbols.Global(owner, D.Parameter(value))
+    s.Scala.Symbols.Global(owner, d.Parameter(value))
   }
 
   def TypeParamSymbol(owner: Symbol, value: String): Symbol = {
-    s.Scala.Symbols.Global(owner, D.TypeParameter(value))
+    s.Scala.Symbols.Global(owner, d.TypeParameter(value))
   }
 
   def SelfSymbol(owner: Symbol): Symbol = {
@@ -59,21 +61,103 @@ trait Symbols {
     s.Scala.Symbols.Multi(List(sym1, sym2))
   }
 
-  implicit class SymbolOps(sym: Symbol) extends wat.MyScalaSymbols.MyScalaSymbolOps(sym) {
-    def companionClass: Symbol = {
-      if (sym.endsWith(".")) sym.substring(0, sym.length - 1) + "#"
-      else NoSymbol
+  private[semantics] class DescriptorParser(s: String) {
+    var i = s.length
+    def fail() = {
+      val message = "invalid symbol format"
+      val caret = " " * i + "^"
+      sys.error(s"$message$EOL$s$EOL$caret")
     }
 
-    def companionObject: Symbol = {
-      if (sym.endsWith("#")) sym.substring(0, sym.length - 1) + "."
-      else NoSymbol
+    val BOF = '\u0000'
+    val EOF = '\u001A'
+    var currChar = EOF
+    def readChar(): Char = {
+      if (i <= 0) {
+        if (i == 0) {
+          i -= 1
+          currChar = BOF
+          currChar
+        } else {
+          fail()
+        }
+      } else {
+        i -= 1
+        currChar = s(i)
+        currChar
+      }
     }
 
-    def companionSymbol: Symbol = {
-      if (sym.endsWith(".")) sym.substring(0, sym.length - 1) + "#"
-      else if (sym.endsWith("#")) sym.substring(0, sym.length - 1) + "."
-      else NoSymbol
+    def parseValue(): String = {
+      if (currChar == '`') {
+        val end = i
+        while (readChar() != '`') {}
+        readChar()
+        s.substring(i + 2, end)
+      } else {
+        val end = i + 1
+        if (!Character.isJavaIdentifierPart(currChar)) fail()
+        while (Character.isJavaIdentifierPart(readChar()) && currChar != BOF) {}
+        s.substring(i + 1, end)
+      }
+    }
+
+    def parseDisambiguator(): String = {
+      val end = i + 1
+      if (currChar != ')') fail()
+      while (readChar() != '(') {}
+      readChar()
+      s.substring(i + 1, end)
+    }
+
+    def parseDescriptor(): d = {
+      if (currChar == '.') {
+        readChar()
+        if (currChar == ')') {
+          val disambiguator = parseDisambiguator()
+          val value = parseValue()
+          d.Method(value, disambiguator)
+        } else {
+          d.Term(parseValue())
+        }
+      } else if (currChar == '#') {
+        readChar()
+        d.Type(parseValue())
+      } else if (currChar == '/') {
+        readChar()
+        d.Package(parseValue())
+      } else if (currChar == ')') {
+        readChar()
+        val value = parseValue()
+        if (currChar != '(') fail()
+        else readChar()
+        d.Parameter(value)
+      } else if (currChar == ']') {
+        readChar()
+        val value = parseValue()
+        if (currChar != '[') fail()
+        else readChar()
+        d.TypeParameter(value)
+      } else {
+        fail()
+      }
+    }
+
+    def entryPoint(): (d, String) = {
+      readChar()
+      val desc = parseDescriptor()
+      (desc, s.substring(0, i + 1))
+    }
+  }
+
+  private[semantics] object DescriptorParser {
+    private val cache = mutable.Map.empty[String, (d, String)]
+
+    def apply(symbol: String): (d, String) = {
+      cache.getOrElseUpdate(symbol, {
+        val parser = new DescriptorParser(symbol)
+        parser.entryPoint()
+      })
     }
   }
 }
